refactor: pull out constants and enums
