--- src/backport/commands.ts
@@ -1,3 +1,0 @@
-export const INIT_REPO = 'INIT_REPO';
-export const SET_UP_REMOTES = 'SET_UP_REMOTES';
-export const BACKPORT = 'BACKPORT';
