--- src/backport/constants.ts
@@ -1,1 +1,0 @@
-export const CHECK_PREFIX = 'Backportable? - ';
