+++ src/constants.ts
@@ -1,0 +1,12 @@
+export const CHECK_PREFIX = 'Backportable? - ';
+
+// trop comment labeling prefixes
+export const TARGET_LABEL_PREFIX = 'target/';
+export const MERGED_LABEL_PREFIX = 'merged/';
+export const IN_FLIGHT_LABEL_PREFIX = 'in-flight/';
+export const NEEDS_MANUAL_LABEL_PREFIX = 'needs-manual-bp/';
+
+// trop repo setup constants
+export const INIT_REPO = 'INIT_REPO';
+export const SET_UP_REMOTES = 'SET_UP_REMOTES';
+export const BACKPORT = 'BACKPORT';
