+++ src/enums.ts
@@ -1,0 +1,4 @@
+export enum PRChange {
+  OPEN,
+  CLOSE,
+}
